COUR-435 Adding frameworks and standards apis a total of 7 more apis. These have stub data for now not backed by cosmos db yet.

+++ src/Dfc.ProviderPortal.ReferenceData/Interfaces/IFeChoice.cs
@@ -6,8 +6,8 @@ namespace Dfc.ProviderPortal.ReferenceData.Interfaces
     {
         Guid Id { get; }
         int UPIN { get; }
-        double LearnerSatisfaction { get; }
-        double EmployerSatisfaction { get; }
+        decimal LearnerSatisfaction { get; }
+        decimal EmployerSatisfaction { get; }
         DateTime? CreatedDateTimeUtc { get; }
     }
 }
